Removed all unused reducers, actions etc
