+++ src/js/actions/ActionTypes.js
@@ -17,29 +17,9 @@ const ActionTypes = {
   ADD_LEXICON_ENTRY: 'ADD_LEXICON_ENTRY',
   // Other actions that need organized
   SET_SAVE_PATH_LOCATION: 'SET_SAVE_PATH_LOCATION',
-  TOGGLE_REMINDER: 'TOGGLE_REMINDER',
-  SET_REMINDER: 'SET_REMINDER',
-  CHANGE_CURRENT_CONTEXT_ID: 'CHANGE_CURRENT_CONTEXT_ID',
-  ADD_COMMENT: 'ADD_COMMENT',
-  ADD_VERSE_EDIT: 'ADD_VERSE_EDIT',
-  ADD_GROUP_DATA: 'ADD_GROUP_DATA',
-  CHANGE_SELECTIONS: 'CHANGE_SELECTIONS',
   STORE_MANIFEST: 'STORE_MANIFEST',
   STORE_PROJECT_SETTINGS: 'STORE_PROJECT_SETTINGS',
   STORE_SAVE_LOCATION: 'STORE_SAVE_LOCATION',
-  LOAD_GROUPS_INDEX: 'LOAD_GROUPS_INDEX',
-  UPDATE_REFRESH_COUNT_GROUPS_INDEX: 'UPDATE_REFRESH_COUNT_GROUPS_INDEX',
-  LOAD_GROUPS_DATA_FROM_FS: 'LOAD_GROUPS_DATA_FROM_FS',
-  SET_REMINDERS_IN_GROUPDATA: 'SET_REMINDERS_IN_GROUPDATA',
-  SET_INVALIDATION_IN_GROUPDATA: 'SET_INVALIDATION_IN_GROUPDATA',
-  TOGGLE_REMINDERS_IN_GROUPDATA: 'TOGGLE_REMINDERS_IN_GROUPDATA',
-  TOGGLE_SELECTIONS_IN_GROUPDATA: 'TOGGLE_SELECTIONS_IN_GROUPDATA',
-  TOGGLE_VERSE_EDITS_IN_GROUPDATA: 'TOGGLE_VERSE_EDITS_IN_GROUPDATA',
-  TOGGLE_MULTIPLE_VERSE_EDITS_IN_GROUPDATA: 'TOGGLE_MULTIPLE_VERSE_EDITS_IN_GROUPDATA',
-  TOGGLE_COMMENTS_IN_GROUPDATA: 'TOGGLE_COMMENTS_IN_GROUPDATA',
-  CLEAR_PREVIOUS_GROUPS_DATA: 'CLEAR_PREVIOUS_GROUPS_DATA',
-  CLEAR_PREVIOUS_GROUPS_INDEX: 'CLEAR_PREVIOUS_GROUPS_INDEX',
-  CLEAR_CONTEXT_ID: 'CLEAR_CONTEXT_ID',
   UPDATE_TARGET_VERSE: 'UPDATE_TARGET_VERSE',
   SHOW_LOADING_CIRCLE: 'SHOW_LOADING_CIRCLE',
   HIDE_LOADING_CIRCLE: 'HIDE_LOADING_CIRCLE',
@@ -98,11 +78,6 @@ const ActionTypes = {
   ADD_PROJECT_SETTINGS_PROPERTY: 'ADD_PROJECT_SETTINGS_PROPERTY',
   SET_PROJECT_TYPE: 'SET_PROJECT_TYPE',
   OLD_SELECTED_PROJECT_FILENAME: 'OLD_SELECTED_PROJECT_FILENAME',
-  // Group menu
-  TOGGLE_MENU_DRAWER: 'TOGGLE_MENU_DRAWER',
-  GROUP_MENU_EXPAND_SUBMENU: 'GROUP_MENU_EXPAND_SUBMENU',
-  GROUP_MENU_SET_FILTER: 'GROUP_MENU_TOGGLE_FILTER',
-  CLEAR_PREVIOUS_FILTERS: 'CLEAR_PREVIOUS_FILTERS',
   // copy right check reducer
   SELECT_PROJECT_LICENSE_ID: 'SELECT_PROJECT_LICENSE_ID',
   LOAD_PROJECT_LICENSE_MARKDOWN: 'LOAD_PROJECT_LICENSE_MARKDOWN',
@@ -136,11 +111,6 @@ const ActionTypes = {
   UPDATE_TOOL_SETTINGS: 'UPDATE_TOOL_SETTINGS',
   SET_SETTING: 'SET_SETTING',
   TOGGLE_SETTING: 'TOGGLE_SETTING',
-  // invalidated reducer
-  SET_INVALIDATED: 'SET_INVALIDATED',
-  SET_INVALIDATED_CHECKS_TOTAL: 'SET_INVALIDATED_CHECKS_TOTAL',
-  SET_VERSE_EDITS_TOTAL: 'SET_VERSE_EDITS_TOTAL',
-  SET_INVALIDATED_ALIGNMENTS_TOTAL: 'SET_INVALIDATED_ALIGNMENTS_TOTAL',
   ADD_TOOL: 'ADD_TOOL',
   OPEN_TOOL: 'OPEN_TOOL',
   CLOSE_TOOL: 'CLOSE_TOOL',
