--- src/js/actions/SideBarActions.js
@@ -1,5 +1,0 @@
-import consts from './ActionTypes';
-
-export function toggleMenu() {
-  return { type: consts.TOGGLE_MENU_DRAWER };
-}
