+++ src/js/components/home/toolsManagement/ToolsCards.js
@@ -21,7 +21,6 @@ const ToolsCards = ({
   loggedInUser,
   projectSaveLocation,
   manifest,
-  // invalidatedReducer,
   toolsCategories,
   originalLanguageBookManifest,
   onMissingResource,
