+++ src/js/containers/home/ToolsManagementContainer.js
@@ -61,7 +61,6 @@ class ToolsManagementContainer extends Component {
           projectSaveLocation,
           toolsCategories,
         },
-        // invalidatedReducer,
       },
       translate,
       originalLanguageBookManifest,
@@ -106,7 +105,6 @@ class ToolsManagementContainer extends Component {
             actions={{ ...this.props.actions }}
             onMissingResource={onMissingResource}
             originalLanguageBookManifest={originalLanguageBookManifest}
-            // invalidatedReducer={invalidatedReducer}
             projectSaveLocation={projectSaveLocation}
             sourceContentUpdateCount={sourceContentUpdateCount}
           />
@@ -125,7 +123,6 @@ const mapStateToProps = (state) => ({
     homeScreenReducer: state.homeScreenReducer,
     projectDetailsReducer: state.projectDetailsReducer,
     loginReducer: state.loginReducer,
-    // invalidatedReducer: state.invalidatedReducer,
   },
 });
 
@@ -164,7 +161,6 @@ ToolsManagementContainer.propTypes = {
   reducers: PropTypes.shape({
     projectDetailsReducer: PropTypes.object.isRequired,
     loginReducer: PropTypes.shape({ loggedInUser: PropTypes.bool }).isRequired,
-    // invalidatedReducer: PropTypes.object.isRequired,
   }).isRequired,
   actions: PropTypes.object.isRequired,
   translate: PropTypes.func.isRequired,
