--- src/js/reducers/contextIdReducer.js
@@ -1,23 +1,0 @@
-import consts from '../actions/ActionTypes';
-
-const initialState = { contextId: null };
-
-const contextIdReducer = (state = initialState, action) => {
-  switch (action.type) {
-  case consts.CHANGE_CURRENT_CONTEXT_ID:
-    return Object.assign({}, state, { contextId: action.contextId });
-  case consts.CLEAR_CONTEXT_ID:
-    return initialState;
-  default:
-    return state;
-  }
-};
-
-export default contextIdReducer;
-
-/**
- * Returns the current context id
- * @param state
- * @return {*}
- */
-export const getContext = (state) => state.contextId;
